+++ CHANGELOG.md
@@ -7,6 +7,12 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.6.3 - Unreleased
 
+### Fixed
+* Issue in `coco_subset` CLI when copy-assets is on and the input is absolute paths.
+
+### Changed
+* `kwcoco validate` now reports the path to the first missing / corrupted image in the error.
+
 
 ## Version 0.6.2 - Released 2023-05-02